+++ examples/Graph embedding demos with PyKEEN/Node embeddings in downstream tasks.lynxkite.json
@@ -19777,7 +19777,7 @@
           ],
           "type": "graph_creation_view"
         },
-        "op_id": "Organize",
+        "op_id": "Table operations > Merge",
         "params": {
           "relations": "{}"
         },
