+++ lynxkite-graph-analytics/src/lynxkite_graph_analytics/operations/__init__.py
@@ -7,3 +7,4 @@ from . import ml_ops  # noqa (imported to trigger registration)
 from . import pykeen_ops  # noqa (imported to trigger registration)
 from . import query_ops  # noqa (imported to trigger registration)
 from . import visualization_ops  # noqa (imported to trigger registration)
+from . import table_ops  # noqa (imported to trigger registration)
