+++ lynxkite-graph-analytics/src/lynxkite_graph_analytics/operations/graph_ops.py
@@ -1,15 +1,100 @@
 """Operations for graphs."""
 
+import json
+import typing
 from lynxkite_core import ops
 from collections import deque
 import pandas as pd
-from .. import core
+from .. import core, bundle
 import networkx as nx
 import enum
 import numpy as np
 
 
-op = ops.op_registration(core.ENV, "Graph operations")
+op = ops.op_registration(core.ENV, "Graph operations", color="green")
+
+
+@op("Merge", icon="link")
+def merge(
+    bundles: list[core.Bundle],
+    *,
+    merge_mode: bundle.BundleMergeMode = bundle.BundleMergeMode.must_be_unique,
+):
+    """Merge multiple inputs"""
+    b = bundle.merge_bundles(bundles, merge_mode=merge_mode)
+    return b
+
+
+@op("Define Edges", view="graph_creation_view", outputs=["output"], icon="link")
+def define_edges(b: core.Bundle, *, relations: str = ""):
+    """Define edges between node tables"""
+    if relations.strip():
+        b.relations = [core.RelationDefinition(**r) for r in json.loads(relations).values()]
+    return ops.Result(output=b, display=b.to_table_view(limit=100))
+
+
+ColumnNameForSource = typing.Annotated[
+    str, {"format": "dropdown", "metadata_query": "[].dataframes[].<source_table>.columns[]"}
+]
+ColumnNameForTarget = typing.Annotated[
+    str, {"format": "dropdown", "metadata_query": "[].dataframes[].<target_table>.columns[]"}
+]
+
+
+@op("Connect nodes on attribute", icon="link")
+def connect_nodes(
+    b: core.Bundle,
+    *,
+    source_table: core.TableName,
+    source_id: ColumnNameForSource,
+    source_attribute: ColumnNameForSource,
+    target_table: core.TableName,
+    target_id: ColumnNameForTarget,
+    target_attribute: ColumnNameForTarget,
+) -> core.Bundle:
+    """
+    Creates edges between nodes from table1 and table2 if the two attributes of the node are equal.
+
+    Parameters:
+    - source_table: Name of the first table
+    - source_id: ID column in the first table
+    - source_attribute: Attribute column in the first table used for matching
+    - target_table: Name of the second table
+    - target_id: ID column in the second table
+    - target_attribute: Attribute column in the second table used for matching
+    """
+    b = b.copy()
+
+    df1 = (b.dfs[source_table]).add_suffix("_src")
+    df2 = (b.dfs[target_table]).add_suffix("_dst")
+    source_key, target_key = source_id, target_id
+    source_attribute, target_attribute, source_id, target_id = (
+        source_attribute + "_src",
+        target_attribute + "_dst",
+        source_id + "_src",
+        target_id + "_dst",
+    )
+    edges = pd.merge(df1, df2, left_on=source_attribute, right_on=target_attribute)
+
+    if source_table == target_table:
+        edges[[source_id, target_id]] = np.sort(edges[[source_id, target_id]], axis=1)
+        edges = edges[edges[source_id] != edges[target_id]].drop_duplicates()
+
+    b.dfs["edges"] = edges
+
+    b.relations.append(
+        core.RelationDefinition(
+            name="graph",
+            df="edges",
+            source_column=source_id,
+            source_table=source_table,
+            source_key=source_key,
+            target_column=target_id,
+            target_table=target_table,
+            target_key=target_key,
+        )
+    )
+    return b
 
 
 @op("Discard loop edges", icon="filter-filled")
@@ -40,7 +125,7 @@ def sample_graph(graph: nx.Graph, *, nodes: int = 100):
     return nx.Graph(graph.subgraph(sample))
 
 
-@op("Graph from edge list", color="green", icon="topology-star-3")
+@op("Graph from edge list", icon="topology-star-3")
 def graph_from_edge_list(
     df: pd.DataFrame, *, source: core.RecordsColumn, target: core.RecordsColumn
 ) -> core.Bundle:
@@ -68,7 +153,7 @@ class DegreeType(enum.StrEnum):
     degree = "degree"
 
 
-@op("NetworkX", "Degree", icon="topology-star-3")
+@op("Degree", icon="topology-star-3")
 def degree(g: nx.Graph) -> nx.Graph:
     g = g.copy()
     nx.set_node_attributes(g, name="in_degree", values=dict(g.in_degree()))
